test - sessão: Unitários

+++ src/main/java/com/desafio/votacao/dtos/sessao_votacao/NovaSessaoDto.java
@@ -2,7 +2,7 @@ package com.desafio.votacao.dtos.sessao_votacao;
 
 public record NovaSessaoDto(
 
-        Long tempo
+        int tempo
 
 ) {
 }
